+++ src/app/feedback/feedback.component.css
@@ -1,0 +1,178 @@
+/* body {
+    font-family: Arial, sans-serif;
+    margin: 0;
+    padding: 0;
+}
+
+.container {
+    width: 80%;
+    margin: 0 auto;
+}
+
+header {
+    background-color: #ff4500;
+    color: #fff;
+    padding: 1em 0;
+    text-align: center;
+}
+
+main {
+    display: flex;
+    justify-content: space-between;
+    padding: 20px 0;
+}
+
+.content {
+    flex: 1;
+}
+
+.post {
+    border: 1px solid #ddd;
+    padding: 10px;
+    margin-bottom: 20px;
+}
+
+.sidebar {
+    width: 25%;
+}
+
+footer {
+    background-color: #f0f0f0;
+    padding: 10px 0;
+    text-align: center;
+} */
+
+body {
+    font-family: Arial, sans-serif;
+    background-color: #f9f9f9;
+    margin: 0;
+    padding: 0;
+}
+
+.home {
+    padding-top: 82px;
+    display: flex;
+    justify-content: center;
+}
+
+.card {
+    width: 100%;
+    background-color: #ffffff;
+    border-radius: 10px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    overflow: hidden;
+}
+
+.card-header {
+    display: flex;
+    align-items: center;
+    padding: 10px;
+    border-bottom: 1px solid #e0e0e0;
+}
+
+.profile-pic {
+    width: 50px;
+    height: 50px;
+    border-radius: 50%;
+    margin-right: 10px;
+}
+
+.user-info p {
+    font-weight: bold;
+    margin: 0;
+}
+
+.user-info span {
+    font-size: 12px;
+    color: #888;
+}
+
+.post-image {
+    width: 100%;
+    height: auto;
+}
+
+.post-text {
+    padding: 10px;
+    font-size: 14px;
+    line-height: 1.4;
+}
+
+.reactions {
+    display: flex;
+    align-items: center;
+    padding: 10px;
+}
+
+.reaction-count {
+    margin-right: 5px;
+    font-weight: bold;
+}
+
+.reaction-icon {
+    width: 20px;
+    height: 20px;
+    margin-right: 5px;
+}
+
+.button {
+    display: inline-block;
+    border: none;
+    background: none;
+    padding: 0;
+    /* margin: 10px; */
+  }
+  .rounded-square-button {
+    background-color: transparent; /* Transparent background */
+    border: 2px solid navy; /* Navy blue border */
+    border-radius: 8px; /* Rounded corners */
+    color: navy; /* Text color */
+    padding: 10px 20px; /* Padding */
+    font-size: 16px; /* Font size */
+    cursor: pointer;
+    outline: none; /* Remove focus outline */
+  }
+  
+  .rounded-square-button:hover {
+    background-color: navy; /* Change background color on hover */
+    color: white; /* Change text color on hover */
+  }
+
+  body {
+    font-family: Arial, sans-serif;
+    background-color: #f9f9f9;
+    margin: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+}
+
+.event-card {
+    border: 1px solid #ddd;
+    padding: 20px;
+    width: 300px;
+    background-color: #fff;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    border-radius: 4px;
+}
+
+
+.event-info p {
+    font-size: 14px;
+    color: #777;
+}
+
+.ticket-info button,
+.ticket-info input[type="text"] {
+    font-size: 16px;
+    padding: 5px 10px;
+    border: none;
+    background-color: #f5f5f5;
+    color: #333;
+    cursor: pointer;
+}
+
+.ticket-info button:hover {
+    background-color: #ddd;
+}
